Moved `ExportCache` to `SystemWorld`
